Added clarity about Number.isNaN functionality
For those that have been working with global isNaN for a while, it can be a bit jarring to have a function with the same name work differently. I added a line to help better describe it's exact function. I also fixed the example given on line 6. It was not correct.